Extensible approach begins
The idea is that _drivers_ bear most of the proof obligations. In order
to achieve extensibility, little fragments of SQL are given as singleton
datatypes like VALUES, TABLE, LIMIT, OFFSET, whatever, and it's up to
the drivers to give instances for these types indicating that they can,
in some cases, be interpreted.

We give one example in this commit in which the postgres universe can
handle insertions of values into tables.