Table/schema creation terms
The idea is that constructing tables and schemas should be done in the
same way as running queries. The only difference: we'll have some way to
generate the DB construction queries from the DB types themselves,
giving the programmer a simple way to lay down her schema in the RDBMS
of choice.